Flyttet kode som kun brukes i test til egen test-klasse
